fix z3 sext
In the definition of function Z3SExt(ctx, bv, wid), the argument(wid) represents the number of bits to be extended. 

So, when calling the function, we should pass the difference between bits to be extended and the current variable's bit-width.
Please refer to ExprOpZExt::GetZ3Expr, which has done this correctly.